requirements.txt
pytest
requests
pyyaml
jsonschema
pytest-html
allure-pytest
pytest-xdist
genson
